Safari-611.1.10/LayoutTests/platform/gtk/http/tests/security/video-cross-origin-accessfailure-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]
EVENT(error)
END OF TEST
